:bug: Skip Docker cache clear if daemon offline

+++ src/clear-cache.ps1
@@ -21,7 +21,7 @@ Options:
 
 # Print ClearCache version string.
 Function Version() {
-    Write-Output 'ClearCache 0.1.1'
+    Write-Output 'ClearCache 0.1.2'
 }
 
 # Script entrypoint.
@@ -48,8 +48,12 @@ Function Main() {
         scoop cache rm --all
     }
 
+    # Check if Docker client is install and Docker daemon is up and running.
     If (Get-Command docker -ErrorAction SilentlyContinue) {
-        docker system prune --force --volumes
+        docker ps 2>&1 | Out-Null
+        If (-Not $LastExitCode) {
+            docker system prune --force --volumes
+        }
     }
 
     If (Get-Command npm -ErrorAction SilentlyContinue) {
